+++ Entities/Item/FairyJar/FairyJar.cs
@@ -138,12 +138,12 @@ namespace FastDragon
 
             Player.ProcessMode = ProcessMode;
 
+            // Keep the music playing, but make it quieter
+            MusicDuckTrigger.Playing = paused;
+
             var musicPlayer = GetTree().FindNode<BackgroundMusicPlayer>();
             if (musicPlayer != null)
-            {
                 musicPlayer.ProcessMode = ProcessMode;
-                MusicDuckTrigger.Playing = paused;
-            }
         }
 
         private string GenerateSaveKey()